==> 03_session.Rmd <==
---
title: "session 3 notes"
author: "Tim Riffe"
date: "2024-07-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Standardization

Take the CDR as our _straw man_:

$$ CDR = \frac{P_x \cdot m_x}{\sum P_x}$$
 
 Where P is population and m is mortality rates
 
 $$ CDR = \sum C_x \cdot m_x$$
 here C is structure, summing to 1.
 
The problem with using $C_x$ observed in a moment to weight together mortality rates is that it is a legacy of past mortality, fertility, and migration, and the relative components thereof are difficult to separate, and in any case they are in the past. The mortality on the other hand is in the present. We would like to compare the mortality in a given moment, separated from its past.

Different kinds of comparisons:
1. If comparing populations in a given period, then it's justified to choose a standard that is somehow in the middle of all the $C_x$ values that are observed. For example, just take the mean, or the overall aggregate.

But, shared features of the individual population structures, such as particularly large cohorts, will move over time (thinking ageing populations), and so comparisons over time will become less usable.

2. comparing over time is quite tricky, because you can't have any cohort features getting in the way. So you'd need to establish a shared standard. For this, it might make sense to use the average survival function covering the whole period, but beware that this will down-weight older ages. A third possibility would be to modify the survival curve according to some population growth rate.


# example

```{r, warning = FALSE, message = FALSE}
library(tidyverse)
dat <- read_csv("https://github.com/timriffe/BSSD2024Module2/raw/master/data/JPNTWN.csv", show_col_types = FALSE)
```

take a look: We would conclude that mortality has been getting worse in both countries, but we know it's not true.
```{r}
dat |> 
  filter(sex == "total") |> 
  group_by(country, year) |> 
  mutate(Cx = exposure / sum(exposure)) |> 
  summarize(CDR = 1000 * sum(Cx * mx, na.rm = TRUE)) |> 
  ggplot(aes(x = year, y = CDR, color = country)) +
  geom_line() +
  theme_minimal()
```

We know a lot here is because of age structure, so we could just swap age structures and see what happens:

```{r}
options(scipen = 2)
dat |> 
  filter(sex == "total") |> 
  group_by(country, year) |> 
  mutate(Cx = exposure / sum(exposure)) |> 
  select( -exposure) |> 
  pivot_wider(names_from = country,
              values_from = c(Cx, mx)) |> 
  summarize(CDR_j = sum(Cx_Japan * mx_Japan, na.rm = TRUE),
            CDR_t = sum(Cx_Taiwan * mx_Taiwan, na.rm = TRUE),
            CDR_jt = sum(Cx_Taiwan * mx_Japan, na.rm = TRUE),
            CDR_tj = sum(Cx_Japan * mx_Taiwan, na.rm = TRUE)) |> 
  filter(year < 2022) |> 
  pivot_longer(-year, 
               names_to = "variant", 
               values_to = "CDR", 
               names_prefix = "CDR_") |> 
  ggplot(aes(x = year, y = CDR, color = variant)) +
  geom_line() +
  theme_minimal() +
  annotate("text", 1990, .011, 
           label = "Taiwan CDR w Japan's structure", angle = 25) +
  annotate("text", 1990, .0074, 
           label = "Japan CDR", angle = 20)+
  annotate("text", 2000, .0062, 
           label = "Taiwan CDR", angle = 8)+
  annotate("text", 2000, .0042, 
           label = "Japan CDR with Taiwan's age structure", angle = 8) +
  guides(color = "none")
```

Notes: at a given point in time, the points are interesting to compare, but note you have to make multiple comparisons to learn something. The time series still gives the same trend due to ageing populations.

```{r}
dat |> 
  filter(sex == "total") |> 
  group_by(country, year) |> 
  mutate(Cx = exposure / sum(exposure)) |> 
  # ggplot(aes(x = age, y = Cx, color = country, group = interaction(country, year), alpha = year)) +
  # geom_line()
  ungroup() |> 
  group_by(age) |> 
  mutate(Cx = mean(Cx, na.rm = TRUE)) |> 
  ungroup() |> 
  # ggplot(aes(x = age, y = Cx)) +
  # geom_line()
  group_by(year, country) |> 
  summarize(ASDR = sum(Cx * mx, na.rm = TRUE),
            .groups = "drop") |> 
  ggplot(aes(x = year, y = ASDR, color = country)) +
  geom_line() +
  theme_minimal()
```
Now we see long term decrease in mortality rates. At least the direction of change is as we expect, but we still might not be so certain about the speed, the level, or much else.

# Decomposition

Let's set up decomposition in general terms before showing examples: $I$ is the thing we calculate, and $\theta$ is everything we need to calculate it. $f()$ is the function that maps parameters to the index (i.e. our calculation method)
$$ I = f(\theta)$$

We denote the difference in the index we calculate (e.g. CDR) using $\Delta$
$$ \Delta = I^2 - I^1$$

Here, $\delta$ is the contribution from each parameter difference to the observed difference in $I$, such that they add up to $\Delta$
$$ \delta = \mathcal{D}(f(), \theta^2, \theta^1)$$

Fancy D is a generic name for a decomposition method. Decompositions should be additive in the sense that:
$$ \Delta = \sum \delta$$

# Kitagawa
A Kitagawa decomposition break apart a difference observed in a weighted mean into a component due to differences in structure (weights, C) and rates (the thing being weighted, m):

$$ \Delta = \Delta^{C} + \Delta^{m}$$

The structure effect $\Delta^C$ is the difference in structure (weights) times the average mortality rates (thing being weighted).

The _rate_ effect (thing being weighted) is the difference in rates times the average structure.

$$ \Delta^C = \sum (C_x^2 - C_x^1) \cdot \overline m_x$$ 
$$ \Delta^m = \sum (m_x^2 - m_x^1) \cdot \overline{C_x}$$
Once the data are in shape, this can be computed in 2 lines in `mutate()`
```{r}
kit <- 
  dat |> 
  filter(year == 2000, 
         sex == "total") |> 
  # create structure
  group_by(country) |> 
  mutate(Cx = exposure / sum(exposure),
         country = substr(country,1,1) |> tolower()) |> 
  # get needed columns side-by-side
  select(-exposure) |> 
  pivot_wider(names_from = country,
              values_from = c(mx, Cx)) |> 
  
  # This is the Kitagawa method:
  mutate(st_effect = (Cx_j - Cx_t) * (mx_t + mx_j) / 2,
         mx_effect = (mx_j - mx_t) * (Cx_t + Cx_j) / 2)
```

Again the formulas to compare with the code:
$$ \Delta^C = \sum (C_x^2 - C_x^1) \cdot \overline m_x$$
$$ \Delta^m = \sum (m_x^2 - m_x^1) \cdot \overline{C_x}$$
Check the sums:
```{r}
kit |> 
  summarize(st_effect = 1000 * sum(st_effect, na.rm = TRUE),
            mx_effect = 1000 * sum(mx_effect, na.rm = TRUE),
            CDR_j = 1000 * sum(Cx_j * mx_j, na.rm = TRUE),
            CDR_t = 1000 * sum(Cx_t * mx_t, na.rm = TRUE),
            CDR_diff = CDR_j - CDR_t) |> 
  mutate(check = st_effect + mx_effect)
```

Age patterns
```{r}
kit |> 
  ggplot(aes(x = age, y = mx_effect)) +
  geom_line()
```

Note: I think the only age pattern here that is worth directly interpreting is the one for mortality rates, not the one for age structure. For age structure the interesting thing is the sum. At least this is my own tentative conclusion.

# Another kind of decomposition

Again, refer to the general expression for a decomposition, where D standard for our chosen general method, f for the function we use to calulate the index I, and the two thetas are our two versions of parameters.
$$ \delta = \mathcal{D}(f(), \theta^2, \theta^1)$$

You function should follow this form
```{r}
my_function <- function(theta){
  # step 1
  # unpack theta into usable pieces
  
  # step 2
  # calculate the index using those pieces
  
  # 3 return the index
}
```

Usage of one of the general decomposition functions follows this general form:
```{r, eval = FALSE}
library(DemoDecomp)
horiuchi(func = my_function,
         pars1 = c(big, fat, vector, of, parameters, first, version),
         pars2 = c(big, fat, vector, of, parameters, second, version))
```


This function takes a single vector of mx and converts it to an approximate life expectancy at birth, assuming we have continuous data. For single age data, the error is quite small.
```{r}
mx_to_e0_quick <- function(mx){
  sum(exp(-cumsum(c(0,mx)))) - .5
}
# dat |> 
#   filter(sex == "total",
#          year == 2000,
#          country == "Japan") |> 
#   pull(mx) |> 
#   mx_to_e0_quick()

# dat |> 
#   filter(sex == "total",
#          year == 2000,
#          country == "Japan") |> 
#   LT_tidy() |> 
#   filter(age == 0) |> 
#   pull(ex)

```

# usage:
Let's compare $e_0$ between Japan and Taiwan in 2000. To use one og the generalized functions inside `mutate()`, you need to manage to stack the values of $\theta^1$ and $\theta^2$ in columns next to each other

```{r}
dec <- 
  dat |> 
  mutate(mx = if_else(is.na(mx),.5,mx)) |> 
  # get parameters side-by-side
  select(-exposure) |> 
  mutate(country = substr(country, 1, 1) |> tolower()) |> 
  pivot_wider(names_from = country, values_from = mx) |> 
  # calculate the decomposition:
  group_by(year,sex) |> 
  mutate(delta = horiuchi(mx_to_e0_quick, `t`, j, N = 10))
```

Check the sum of delta: should equal the life expectancy difference with a small residual in this case.
```{r}
dec |> 
  group_by(year, sex) |> 
  summarize(Delta = sum(delta))
```

View the age pattern of the decomposition elements:
```{r}
dec |> 
  filter(year %in% c(1970, 2019),
         sex != "total") |> 
  ggplot(aes(x = age, y = delta, color = sex, linetype = as.factor(year))) +
  geom_line()
```
